requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.7
django-countries==7.0
django-dark==0.1.9
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-mptt==0.11.0
django-mptt-admin==2.1.0
django-tastypie==0.14.3
flake8==3.8.4
marshmallow==3.11.1
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.4
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2021.1
rstr==2.2.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
